Integrate audio normalization into transcode. Apply it only when exactly 1 video exists for the stream.
